Improve PoolAllocatorTFH alignment and add static asserts
Refactored PoolAllocatorTFH to use static_asserts for alignment and size requirements, and added platform-specific handling for alignment calculation. Updated PoolMinAlignment to 4 for broader compatibility. Fixed macro typo in LogMacros.h and included missing headers in several files. Adjusted MoveTest in tests for type consistency.

+++ tests/PoolAllocatorTFH_tests.cpp
@@ -15,10 +15,10 @@ namespace {
     struct MoveTest {
         static std::atomic<int> constructions;
         static std::atomic<int> destructions;
-        int value;
+        size_t value;
         char padding[sizeof(std::size_t)]; // Ensure object is large enough for freelist
 
-        MoveTest(int val) : value(val) { ++constructions; }
+        MoveTest(size_t val) : value(val) { ++constructions; }
         MoveTest(const MoveTest&) = delete;
         MoveTest(MoveTest&& other) noexcept : value(other.value) { other.value = -1; ++constructions; }
         ~MoveTest() { ++destructions; }
